fix: use BigDecimal instead of double for prices

+++ src/main/java/model/order/Order.java
@@ -2,12 +2,13 @@ package model.order;
 
 import model.restaurant.Restaurant;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public interface Order {
     LocalDate getDate();
 
-    Double getPrice();
+    BigDecimal getPrice();
 
     boolean involvesRestaurant(Restaurant restaurant);
 }
